@@ -1,16 +1,13 @@
-using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
-using System.Threading.Tasks;
 
-namespace Expense.Models
+namespace expense.Models;
+
+public interface IExpenseContext
 {
-    public interface IExpenseContext
-    {
-          Task<ActionResult<IEnumerable<ExpenseItem>>> ListAsync();
-          Task<ActionResult<IEnumerable<ExpenseItem>>> ListAsyncByTripId(string tripId);
-          Task<ExpenseItem> GetExpense(string id);
-          Task<int> AddExpenseItem(ExpenseItem item);
-          Task<int> UpdateExpenseItem(ExpenseItem item);
-          Task<int> DeleteExpenseItem(ExpenseItem item);
-    }
+    Task<ActionResult<IEnumerable<ExpenseItem>>> ListAsync();
+    Task<ActionResult<IEnumerable<ExpenseItem>>> ListAsyncByTripId(string tripId);
+    Task<ExpenseItem> GetExpense(string id);
+    Task<int> AddExpenseItem(ExpenseItem item);
+    Task<int> UpdateExpenseItem(ExpenseItem item);
+    Task<int> DeleteExpenseItem(ExpenseItem item);
 }
